fix: 구매금액이 숫자가 아닐시 ERROR 메시지 출력

+++ src/main/java/lotto/service/Statistics.java
@@ -7,7 +7,7 @@ public class Statistics {
 
     private double rate;
     public Statistics(int money, RankResult rankResult) {
-        this.rate = (getTotalMoney(rankResult) / money)*100;
+        this.rate = (getTotalMoney(rankResult) / (double)money)*100;
     }
     private long getTotalMoney(RankResult rankResult) {
         int totalMoney = 0;
